extract quic to a different file
