+++ src/net/src/lib.rs
@@ -1,6 +1,11 @@
 #[macro_use]
 extern crate log;
 
+use self::challenger::Challenger;
+use self::net::ArcFile;
+use self::net::QuicAddr;
+use self::net::Something;
+use self::net::MAX_FRAME_SIZE;
 use self::util::normalize;
 use self::util::peek_quic_varint;
 use self::util::secure_hash;
@@ -17,26 +22,29 @@ macro_rules! bail {
     }
 }
 
+mod challenger;
 mod error;
 mod ffi;
 mod key;
-mod proto;
+mod net;
+mod quic;
 mod util;
 
 pub use self::error::Error;
 pub use self::error::Result;
 pub use self::key::Identity;
 pub use self::key::PrivateIdentity;
-pub use self::proto::Event;
-pub use self::proto::Net;
-pub use self::proto::NetBuilder;
-pub use self::proto::PeerIndex;
+pub use self::net::Event;
+pub use self::net::Net;
+pub use self::net::NetBuilder;
+pub use self::net::PeerIndex;
 
 pub fn client2_main() -> Result<()> {
     env_logger::init();
-    let identity = "6a35f116046280da62602ad979e151daee8e9109befd34337427bbb968c2bcbc"
-        .parse()
-        .unwrap();
+    let identity =
+        "6a35f116046280da62602ad979e151daee8e9109befd34337427bbb968c2bcbc"
+            .parse()
+            .unwrap();
     let remote = "ddnet-15+quic://127.0.0.1:4433#0026a0d653cd5f38d1002bf166167933f2f7910f26d6dd619b2b3fe769e057ee";
     let mut builder = Net::builder();
     builder.identity(identity);
@@ -44,9 +52,11 @@ pub fn client2_main() -> Result<()> {
     let mut buf = [0; 65536];
     let conn_idx = net.connect(&remote).context("Net::connect")?;
     loop {
-        while let Some((idx, event)) = net.recv(&mut buf).context("Net::recv")? {
+        while let Some((idx, event)) =
+            net.recv(&mut buf).context("Net::recv")?
+        {
             assert_eq!(conn_idx, idx);
-            use proto::Event::*;
+            use self::Event::*;
             match event {
                 Connect(id) => {
                     println!("{}: connect {}", idx, id);
@@ -82,9 +92,10 @@ pub fn client2_main() -> Result<()> {
 pub fn server2_main() -> Result<()> {
     env_logger::init();
     let bindaddr = "0.0.0.0:4433".parse().unwrap();
-    let identity = "5cf2a4f0ed3dc85b3f4bfa5ca97b8adeaf0d5ec165179af805a3f7751dd8ac47"
-        .parse()
-        .unwrap();
+    let identity =
+        "5cf2a4f0ed3dc85b3f4bfa5ca97b8adeaf0d5ec165179af805a3f7751dd8ac47"
+            .parse()
+            .unwrap();
     let mut builder = Net::builder();
     builder.identity(identity);
     builder.bindaddr(bindaddr);
@@ -92,8 +103,10 @@ pub fn server2_main() -> Result<()> {
     let mut net = builder.open().context("Net::open")?;
     let mut buf = [0; 65536];
     loop {
-        while let Some((idx, event)) = net.recv(&mut buf).context("Net::recv")? {
-            use proto::Event::*;
+        while let Some((idx, event)) =
+            net.recv(&mut buf).context("Net::recv")?
+        {
+            use self::Event::*;
             match event {
                 Connect(id) => println!("{}: connect {}", idx, id),
                 Chunk(size, nonvital) => {
